version 3.0.1

+++ src/main/java/com/internship/HRapp/service/concretes/AddressService.java
@@ -3,37 +3,52 @@ package com.internship.HRapp.service.concretes;
 import com.internship.HRapp.dto.addressDto.AddressDto;
 import com.internship.HRapp.entity.Address;
 import com.internship.HRapp.mapper.AddressMapper;
+import com.internship.HRapp.mapper.UserMapper;
 import com.internship.HRapp.repository.AddressRepo;
 import com.internship.HRapp.service.interfaces.AddressServiceInterface;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import javax.validation.constraints.NotNull;
 import java.util.List;
 import java.util.UUID;
 
 @Service
+@Transactional
 @RequiredArgsConstructor
-public class AddressServiceImpl implements AddressServiceInterface {
+public class AddressService implements AddressServiceInterface {
     private final AddressRepo addressRepo;
     private final AddressMapper addressMapper;
+    private final UserMapper userMapper;
 
     @Override
-    public AddressDto getAddressById(UUID addressid) {
-        return addressMapper.modeltoDto(addressRepo.getById(addressid));
+    public AddressDto getAddressById(UUID addressID){
+        boolean exists = addressRepo.existsById(addressID);
+        if (!exists){
+            throw new IllegalStateException(
+                    "Address with id " + addressID + " does not exist!"
+            );
+        }
+        return addressMapper.modeltoDto(addressRepo.getById(addressID));
     }
 
     @Override
-    public List<AddressDto> getAddresses() {
+    public List<AddressDto> getAddresses(){
         return addressMapper.toDto(addressRepo.findAll());
     }
 
     @Override
-    public String deleteAddressById(UUID addressID) {
+    public String deleteAddressById(UUID addressID){
+        boolean exists = addressRepo.existsById(addressID);
+        if (!exists){
+            throw new IllegalStateException(
+                    "Address with id " + addressID + " does not exist!"
+            );
+        }
         addressRepo.deleteById(addressID);
         return "address removed {}" + addressID;
     }
-
 
     @Override
     public AddressDto addNewAddress(AddressDto addressDto) {
@@ -42,8 +57,8 @@ public class AddressServiceImpl implements AddressServiceInterface {
     }
 
     @Override
-    public void editAddress(@NotNull AddressDto addressDto) {
-        Address address = addressRepo.getAddressByAddressID(addressDto.getAddressID());
+    public void  editAddress(@NotNull AddressDto addressDto) {
+        Address address =addressRepo.getAddressByAddressID(addressDto.getAddressID());
         address.setCity(addressDto.getCity());
         address.setState(addressDto.getState());
         address.setStreet(addressDto.getStreet());
